Refactor to use non-synchronized FastByteArrayOutputStream
